chattrsnoop: fix immutable attribute set on s390
On s390 the plugin was reporting when the immutable attribute was
cleared, but was not set. The ioctl_iflags man page says the
argument is an int *.

+++ vql/linux/bpf/chattrsnoop/chattrsnoop.bpf.c
@@ -122,7 +122,7 @@ SEC("kprobe/do_vfs_ioctl")
 int BPF_KPROBE(trace_vfs_ioctl, struct file *filp, unsigned int fd,
 	       unsigned int cmd, unsigned long arg)
 {
-	unsigned long flags;
+	int flags;
 	struct inode *inode = BPF_CORE_READ(filp, f_inode);
 
 	if (cmd != FS_IOC_SETFLAGS)
